mover: allow moves to be committed by specifying a mover type
When creating a mover class, add a template argument that allows the
user to say whether they want the move to be committed or reverted.
If the move is to be committed the destructor of the mover class is
empty, which causes the move that was done in the constructor to be
committed to the board.