search: don't use the result of a partial PV
Since we don't know how far through the search the current PV will be
when the search is interrupted, don't use it.  It could have only
explored one node into the tree, in which case the move selected will
be bad.

Instead, use the PV that was obtained from the previous iteration.