movegen: allow a hashmove to be passed into movegen
Allow a hash move, that is one that was found as a PV node in a
previous search, to be passed into the move list.  If that is a valid
move and is in the current move list, move it to the front of the
movelist, so it is considered first.